requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
redis>=5.0.0
celery>=5.3.0
channels>=4.0.0
channels-redis>=4.1.0
python-decouple>=3.8
Pillow>=10.0.0
web3>=6.0.0
eth-account>=0.10.0
gunicorn>=21.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0